requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
Pillow==8.4.0
pytesseract==0.3.8
Werkzeug==2.0.2
